Refactor code to use CapsulaConfig for CaptureConfig and MonitorConfig
