+++ capsula/capture.py
@@ -1,95 +1,27 @@
 from __future__ import annotations
 
+__all__ = [
+    "capture",
+]
+
 import logging
 import subprocess
-import sys
-
-if sys.version_info < (3, 11):
-    from datetime import timezone as _timezone
-
-    UTC = _timezone.utc
-
-else:
-    from datetime import UTC
-from datetime import datetime
-from pathlib import Path  # noqa: TCH003
-
-from pydantic import BaseModel, ConfigDict, Field
+from typing import TYPE_CHECKING
 
 from capsula.context import Context
-from capsula.file import CaptureFileConfig  # noqa: TCH001 for pydantic
 from capsula.globalvars import set_capsule_dir
+
+if TYPE_CHECKING:
+    from capsula.config import CapsulaConfig
 
 logger = logging.getLogger(__name__)
 
 
-def to_hyphen_case(string: str) -> str:
-    return string.replace("_", "-")
-
-
-class GitConfig(BaseModel):
-    repositories: dict[str, Path] = Field(default_factory=dict)
-
-
-class CaptureConfig(BaseModel):
-    """Configuration for the capture command."""
-
-    model_config = ConfigDict(
-        alias_generator=to_hyphen_case,
-        populate_by_name=True,
-        extra="forbid",
-    )
-
-    vault_directory: Path
-    capsule_template: str
-
-    # Whether to include the Capsula version in the output file.
-    # include_capsula_version: bool = True
-
-    include_cpu: bool = True
-
-    pre_capture_commands: list[str] = Field(default_factory=list)
-
-    environment_variables: list[str] = Field(default_factory=list)
-
-    files: dict[Path, CaptureFileConfig] = Field(default_factory=dict)
-
-    git: GitConfig = Field(default_factory=GitConfig)
-
-    _capsule_directory: Path | None = None
-    _root_directory: Path | None = None
-
-    @property
-    def capsule(self) -> Path:
-        if self._capsule_directory is None:
-            self._capsule_directory = (
-                self.root_directory
-                / self.vault_directory
-                / datetime.now(UTC)
-                .astimezone()
-                .strftime(
-                    self.capsule_template,
-                )
-            )
-        return self._capsule_directory
-
-    @property
-    def root_directory(self) -> Path:
-        if self._root_directory is None:
-            msg = "Project root is not set"
-            raise ValueError(msg)
-        return self._root_directory
-
-    @root_directory.setter
-    def root_directory(self, value: Path) -> None:
-        self._root_directory = value
-
-
-def capture(*, config: CaptureConfig) -> Context:
+def capture(*, config: CapsulaConfig) -> Context:
     """Capture the context."""
-    logger.debug(f"Capture config: {config}")
+    logger.debug(f"Capture config: {config.capture}")
 
-    for command in config.pre_capture_commands:
+    for command in config.capture.pre_capture_commands:
         logger.info(f"Running pre-capture command: {command!r}")
         try:
             result = subprocess.run(
