+++ tests/test_capture.py
@@ -1,18 +1,21 @@
 import tempfile
 from pathlib import Path
 
-from capsula.capture import CaptureConfig, capture
+from capsula.capture import capture
+from capsula.config import CapsulaConfig, CaptureConfig, MonitorConfig
 
 
 def test_capture() -> None:
     with tempfile.TemporaryDirectory() as root_directory:
-        capture_config = CaptureConfig(
+        capsula_config = CapsulaConfig(
             vault_directory=Path(root_directory) / "vault",
             capsule_template=r"%Y%m%d_%H%M%S",
+            capture=CaptureConfig(),
+            monitor=MonitorConfig(),
         )
-        capture_config.root_directory = Path(root_directory)
+        capsula_config.root_directory = Path(root_directory)
 
-        ctx = capture(config=capture_config)
+        ctx = capture(config=capsula_config)
 
         assert ctx.platform is not None
         assert ctx.cpu is not None
